requirements.txt
requests==2.32.3
python-dotenv==1.0.1
playwright==1.52.0
APScheduler==3.10.4
